requirements.txt
et-xmlfile        == 1.1.0
numpy             == 1.26.2
opencv-python     == 4.9.0.80
openpyxl          == 3.1.2
packaging         == 23.2
pandas            == 2.1.4
pdf2image         == 1.17.0
pillow            == 10.2.0
pip               == 23.0.1
PyPDF2            == 3.0.1
pytesseract       == 0.3.10
python-dateutil   == 2.8.2
pytz              == 2023.3.post1
setuptools        == 65.5.0
six               == 1.16.0
typing_extensions == 4.9.0
tzdata            == 2023.3